issue/209 - Migrate from npm to yarn for package management (#214)
* Migrate front-end from npm -> yarn

* Migrate back-end from npm -> yarn

* Replace npm with yarn

This updates the documentation and changes references from npm to yarn.

* Migrate build/deploy/CI with yarn

Note that `yarn install --frozen-lockfile` replaces `npm ci`
and that `--if-present` is dropped from `npm run build` when moving to
`yarn build` as it is only needed when there is no strict control over
the package.json. We have such control, so it is unnecessary.

* Remove lingering file generated by npm

* Update scripts to use yarn

* Fix build command

* Attempt to fix build commands in deployment config

* Ensure no tests are treated as 'pass'

* Update CI workflow to resolve issues

This splits frontend and backend into separate jobs, whilst also trying
to maintain a consistent and meaningful layout for the config.

* Revert back to 'yarn build'

This reverts commit e415bebdff70cc4f1e63d8d9975d54fda94338d0.

* Fix typo; 'runs' to 'run'

* Remove build command from backend

* Change order so build is before test

The frontend tests rely on files generated by the build process, so
it's important to build before running the tests.